🎨 ADD : like post method 추가

+++ src/api/like.tsx
@@ -1,31 +1,37 @@
 import { Context } from "elysia";
 import Like from "../model/like";
 
-export const likeHeart = async (req: any, res: any) => {
-  try {
-    const { id } = req.params;
-    const { isLiked } = req.body;
-    const existingLike = await Like.findOne({ _id: id });
+export const likeHeart = async (c: Context<{ params: { id: string } }>) => {
+  if (c.params && !c.params?.id) {
+    c.set.status = 400;
+    throw new Error("No id provided");
+  }
 
-    if (existingLike) {
-      await Like.deleteOne({ _id: id });
-      console.log("삭제되었습니다. ");
-      res.json({
-        _id: id,
-        like: isLiked,
-        message: "Like deleted successfully.",
-      });
-    } else {
-      const newLike = new Like({ _id: id, like: isLiked });
-      const savedLike = await newLike.save();
-      console.log(newLike);
-      res.json(savedLike);
-    }
-  } catch (error) {
-    console.error("Error adding like:", error);
-    res.status(500).json({ error: "Internal Server Error" });
+  const id = c.params.id;
+  if (!c.body) throw new Error("No body provided");
+  const { isLiked } = c.body;
+
+  const existingLike = await Like.findById(id);
+  if (existingLike) {
+    await Like.deleteOne({ _id: id });
+    return {
+      status: c.set.status,
+      success: true,
+      data: existingLike,
+      message: "Like deleted successfully",
+    };
+  } else {
+    const newLike = new Like({ _id: id, like: isLiked });
+    const savedLike = await newLike.save();
+    return {
+      status: c.set.status,
+      success: true,
+      data: savedLike,
+      message: "Like created successfully",
+    };
   }
 };
+
 export const getHeart = async (c: Context) => {
   const like = await Like.find();
   if (!like || like.length === 0) {
